HU16 se organizaron los metodos correspondientes en admin y client
