+++ src/main/java/com/pragma/powerup/usermicroservice/adapters/driving/http/handlers/IClientHandler.java
@@ -1,7 +1,9 @@
 package com.pragma.powerup.usermicroservice.adapters.driving.http.handlers;
 
 import com.pragma.powerup.usermicroservice.adapters.driving.http.dto.request.CreateClientRequestDto;
+import com.pragma.powerup.usermicroservice.adapters.driving.http.dto.response.UserResponseDto;
 
 public interface IClientHandler {
     void saveClient(CreateClientRequestDto createClientRequestDto);
+    UserResponseDto getClient(Long id);
 }
