+++ src/main/java/com/pragma/powerup/usermicroservice/domain/api/IClientServicePort.java
@@ -4,4 +4,5 @@ import com.pragma.powerup.usermicroservice.domain.model.User;
 
 public interface IClientServicePort {
     void saveClient(User user);
+    User getClient(Long id);
 }
